+++ src/css/Header.css
@@ -26,7 +26,7 @@ li {
 }
 
 .link:hover {
-  background-color: #111;
+  background-color: rgb(0, 17, 95);
 }
 
 .active {
